movies/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
psycopg2==2.9.7
py==1.11.0
pydantic==2.3.0
pydantic-core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
retry==0.9.2
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing-extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
